I deleted the items from family branch
