+++ menu-api/src/items/Family.interface.ts
@@ -1,0 +1,11 @@
+export interface Offspring {
+  name: string;
+}
+
+export interface BaseFamily {
+  name: string;
+}
+
+export interface FamilyItem extends BaseFamily {
+  id: number;
+}
