+++ menu-api/src/items/Family.items.interface.ts
@@ -1,0 +1,7 @@
+// src/items/items.interface.ts
+
+import { FamilyItem, BaseFamily } from "./Family.interface";
+
+export interface FamilyItems {
+  [key: number]: FamilyItem;
+}
