+++ menu-api/src/items/Offspring.service.ts
@@ -1,0 +1,9 @@
+import { Offspring } from "./Family.interface";
+
+let offspring: Offspring = {
+  name: "child1",
+};
+
+export const findAll = async (): Promise<{ offspring: Offspring[] }> => ({
+  offspring: [{ name: "child1" }, { name: "child2" }],
+});
