--- menu-api/src/items/item.interface.ts
@@ -1,14 +1,0 @@
-export interface BaseItem {
-  name: string;
-  price: number;
-  description: string;
-  image: string;
-}
-
-export interface Item extends BaseItem {
-  id: number;
-}
-
-export interface Offspring {
-  name: string;
-}
