--- menu-api/src/items/items.interface.ts
@@ -1,7 +1,0 @@
-// src/items/items.interface.ts
-
-import { Item } from "./item.interface";
-
-export interface Items {
-  [key: number]: Item;
-}
